pyproject.toml
[project]
name = "mcp-server-qdrant"
version = "0.8.0"
description = "MCP server for retrieving context from a Qdrant vector database"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "fastembed>=0.6.0",
    "qdrant-client>=1.12.0",
    "pydantic>=2.10.6",
    "fastmcp>=2.7.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
gemini = ["google-generativeai>=0.8.0"]
all = [
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipdb>=0.13.13",
    "isort>=6.0.1",
    "mypy>=1.9.0",
    "pre-commit>=4.1.0",
    "pyright>=1.1.389",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
]

[project.scripts]
mcp-server-qdrant = "mcp_server_qdrant.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
